pyproject.toml
[project]
name = "research-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "pillow>=10.0.0",
    "tensorflow-datasets>=4.9.9",
    "datasets>=4.4.2",
    "smbprotocol>=1.15.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []  # No Python packages to install, just dependencies
